Function namedlock::arcmutexguard::arc_mutex_lock

pub fn arc_mutex_lock<'a, T>(mutex: Arc<Mutex<T>>) -> LockResult<ArcMutexGuard<'a, T>>

Acquires an Arc<Mutex<_>>, blocking the current thread until it is able to do so.

This function will block the local thread until it is available to acquire the mutex. Upon returning, the thread is the only thread with the mutex held. An RAII guard is returned to allow scoped unlock of the lock. When the guard goes out of scope, the mutex will be unlocked, and the Arc will be unreferenced.