Struct namedlock::LockSpace

pub struct LockSpace<K: Eq + Hash, V> {
    // some fields omitted
}

A LockSpace<K,V> holds many Mutex<V>'s, keyed by K.

All accesses to the internal value must go through the with_lock() method.

See the crate documentation for an example.

Methods

impl<K: Eq + Hash + Clone, V> LockSpace<K, V>

fn new(auto_cleanup: bool) -> LockSpace<K, V>

Create a new LockSpace.

If auto_cleanup is true, values will be deleted automatically when the last lock is released. Otherwise, values will remain in the space until try_remove() returns Success.

fn with_lock<F, C, R, Q: ?Sized + Hash + Eq>(&self, key: &Q, initial: C, f: F) -> R where K: Borrow<Q>, &'a Q: Into<K>, C: FnOnce() -> V, F: FnOnce(&mut V) -> R

Find the object by key, or create it by calling initial if it does not exist. Then, call f on that object.

let space=namedlock::LockSpace::<String,i32>::new(false);

space.with_lock("test",||0,|i|*i+=1);
assert_eq!(space.with_lock("test",||0,|i|*i),1);

fn try_remove<Q: ?Sized + Hash + Eq>(&self, key: &Q) -> LockSpaceRemoveResult where K: Borrow<Q>

Find the object by key, then delete it if it is not actively being used. If it is actually being used, WouldBlock will be returned.