Struct namedlock::LockSpace

pub struct LockSpace<K: Eq + Hash, V> {
    // some fields omitted
}

A LockSpace<K,V> holds many Mutex<V>'s, keyed by K.

All accesses to the internal value must go through one of the lock methods.

See the crate documentation for an example.

Key parameters

Most of the LockSpace<K,V> methods take a key: &Q where K: Borrow<Q>, like a HashMap. This basically means "anything that's like a K." However, keys are stored internally inside an Arc. Not all types that normally have K: Borrow<Q>, have Arc<K>: Borrow<Q>, and Rust currently does not allow one to implement that generically. This module re-implements the special Borrow implementations for String, OsString, PathBuf, Vec, and Cow. If you are using a custom K that has a non-standard K: Borrow<_>, you will need to impl Borrow<_> for ArcHashKey<K>, for example using the macro ahk_chain_borrow!. This macro is used internally like so:

ahk_chain_borrow! {
    String: (),
    OsString: (),
    PathBuf: (),
    Vec<T>: ( T, ),
}

It's possible to specify more than one type parameter this way. Lifetime parameters are not supported.

Methods

impl<K: Eq + Hash + Clone, V> LockSpace<K, V>

fn new(cleanup: Cleanup) -> LockSpace<K, V>

Create a new LockSpace.

If cleanup is AutoCleanup, values will be deleted automatically when the last lock is released. Otherwise, values will remain in the space until try_remove() returns Success.

fn lock<'a, C, Q: ?Sized + Hash + Eq>(&'a self, key: &Q, initial: C) -> LockResult<LockSpaceGuard<'a, K, V>> where ArcHashKey<K>: Borrow<Q>, &'q Q: Into<K>, C: FnOnce() -> V

Find the object by key, or create it by calling initial if it does not exist. Then, lock it and return a LockSpaceGuard over the object. Once the guard is dropped, its object is unlocked, and if AutoCleanup is specified for this space, removed if this is the last use.

The key may be any borrowed form of the map's key type, but see the struct documentation for a note.

let space=namedlock::LockSpace::<String,i32>::new(namedlock::KeepUnused);

let value=space.lock("test",||0);
*value.unwrap()+=1;
let value=space.lock("test",||0);
assert_eq!(*value.unwrap(),1);

fn with_lock<F, R, C, Q: ?Sized + Hash + Eq>(&self, key: &Q, initial: C, f: F) -> LockResult<R> where ArcHashKey<K>: Borrow<Q>, &'q Q: Into<K>, C: FnOnce() -> V, F: FnOnce(&mut V) -> R

Find the object by key, or create it by calling initial if it does not exist. Then, call f on that object.

The key may be any borrowed form of the map's key type, but see the struct documentation for a note.

let space=namedlock::LockSpace::<String,i32>::new(namedlock::KeepUnused);

space.with_lock("test",||0,|i|*i+=1);
assert_eq!(space.with_lock("test",||0,|i|*i).unwrap(),1);

fn try_remove<Q: ?Sized + Hash + Eq>(&self, key: &Q) -> LockSpaceRemoveResult where ArcHashKey<K>: Borrow<Q>

Find the object by key, then delete it if it is not actively being used. If it is actually being used, WouldBlock will be returned.

This is only useful if this LockSpace is of the KeepUnused kind.

The key may be any borrowed form of the map's key type, but see the struct documentation for a note.